chore: change /club/dashboard/reservation -> /shop
