--- src/app/club/dashboard/reservation/host/[id]/page.tsx
@@ -1,15 +1,0 @@
-import BackButton from "@/components/dashboard/common/BackButton";
-import HostOverview from "@/components/dashboard/club/reservation/molecules/HostOverview";
-import HostOtherProduct from "@/components/dashboard/club/reservation/organisms/HostOtherProduct";
-import HostReview from "@/components/dashboard/club/reservation/organisms/HostReview";
-
-export default function HostDetailPage() {
-  return (
-    <>
-      <BackButton />
-      <HostOverview />
-      <HostOtherProduct />
-      <HostReview />
-    </>
-  );
-}
