--- src/app/club/dashboard/reservation/inquiry/page.tsx
@@ -1,11 +1,0 @@
-import InquiryOverview from "@/components/dashboard/club/reservation/molecules/InquiryOverview";
-import InquiryList from "@/components/dashboard/club/reservation/organisms/InquiryList";
-
-export default function ReservationInquiryPage() {
-  return (
-    <>
-      <InquiryOverview />
-      <InquiryList />
-    </>
-  );
-}
