--- src/app/club/dashboard/reservation/item/[id]/inquiry/page.tsx
@@ -1,13 +1,0 @@
-import BackButton from "@/components/dashboard/common/BackButton";
-import InquiryForm from "@/components/dashboard/club/reservation/organisms/InquiryForm";
-import InquirySuccessModal from "@/components/dashboard/club/reservation/organisms/InquirySuccessModal";
-
-export default function ItemInquiryPage() {
-  return (
-    <>
-      <BackButton />
-      <InquiryForm />
-      <InquirySuccessModal />
-    </>
-  );
-}
