+++ src/app/club/dashboard/shop/inquiry/[id]/page.tsx
@@ -1,5 +1,5 @@
 import BackButton from "@/components/dashboard/common/BackButton";
-import InquiryDetail from "@/components/dashboard/club/reservation/organisms/InquiryDetail";
+import InquiryDetail from "@/components/dashboard/club/shop/organisms/InquiryDetail";
 
 export default function InquiryDetailPage() {
   return (
