+++ src/app/club/dashboard/shop/inquiry/page.tsx
@@ -1,0 +1,11 @@
+import InquiryOverview from "@/components/dashboard/club/shop/molecules/InquiryOverview";
+import InquiryList from "@/components/dashboard/club/shop/organisms/InquiryList";
+
+export default function InquiryPage() {
+  return (
+    <>
+      <InquiryOverview />
+      <InquiryList />
+    </>
+  );
+}
