+++ src/app/club/dashboard/shop/item/[id]/inquiry/page.tsx
@@ -1,0 +1,15 @@
+import BackButton from "@/components/dashboard/common/BackButton";
+import InquiryForm from "@/components/dashboard/club/shop/organisms/InquiryForm";
+import InquirySuccessModal from "@/components/dashboard/club/shop/modals/InquirySuccessModal";
+
+export default function ItemInquiryPage() {
+  return (
+    <>
+      <BackButton />
+      <InquiryForm />
+      <div className="m-0">
+        <InquirySuccessModal />
+      </div>
+    </>
+  );
+}
