+++ src/app/club/dashboard/shop/layout.tsx
@@ -1,4 +1,4 @@
-export default function ReservationLayout({
+export default function Layout({
   children,
 }: Readonly<{
   children: React.ReactNode;
