+++ src/components/dashboard/club/shop/organisms/ReservationList.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import Pagination from "@/components/dashboard/common/Pagination";
-import ReservationTable from "@/components/dashboard/club/reservation/molecules/ReservationTable";
+import ReservationTable from "@/components/dashboard/club/shop/molecules/ReservationTable";
 
 export default function ReservationList() {
   const [currentPage, setCurrentPage] = useState<number>(1);
